add work around for swmp build bug
