+++ README.md
@@ -1,6 +1,6 @@
 # 🦙🌲🤏 Alpaca-LoRA && Starwhale
 
-This repo is forked from [tloen/alpaca-lora](https://github.com/tloen/alpaca-lora), see the [original README.md](README_original.md). We add some Starwhale support for this repo, users could manage lifecycle of the model by starwhale, including:
+This repo is forked from [tloen/alpaca-lora](https://github.com/tloen/alpaca-lora), see the [original README.md](README_original.md). We add some Starwhale support for this repo, users could manage lifecycle of the model/dataset by starwhale, including:
 - finetune a new version of model locally or remotelly and get a finetuned version of model
 - serve an API locally or remotelly
 - evaluate the model with Starwhale datasets
